ajout middleware

+++ back-end/src/Controller/DeliveryController.php
@@ -8,6 +8,7 @@ use App\Entity\DeliveryAddress;
 use App\Entity\User;
 use App\Entity\Vehicle;
 use App\Entity\VehicleType;
+use App\Middleware\AuthentificationMiddleware;
 use DateInterval;
 use DateTime;
 use Doctrine\Persistence\ManagerRegistry;
@@ -23,6 +24,10 @@ use Symfony\Component\Security\Http\Attribute\CurrentUser;
 #[Route('/api/delivery', name: 'api_')]
 class DeliveryController extends AbstractController
 {
+    public function __construct(private AuthentificationMiddleware $authentificationMiddleware)
+    {
+    }
+
     const UPCOMING = 0;
     const STATUS_PENDING = 1;
     const STATUS_END = 2;
@@ -30,6 +35,11 @@ class DeliveryController extends AbstractController
     #[Route('/add', name: 'delivery_create', methods: ['post'])]
     public function create(ManagerRegistry $doctrine, Request $request, #[CurrentUser] ?User $user): JsonResponse
     {
+        if (!$this->authentificationMiddleware->checkIfUserDriver($request)) {
+            return $this->json([
+                'message' => 'You are not authentified or doesn\'t have the right to access this page'
+            ], Response::HTTP_UNAUTHORIZED);
+        }
         $dotenv = new Dotenv();
         $dotenv->load('../.env');
         $geoApiUrl = 'https://api.opencagedata.com/geocode/v1/json?key=d030bd453deb4ee0b589b32474216bf0&q=';
@@ -143,6 +153,11 @@ class DeliveryController extends AbstractController
     #[Route('/get/{id}', name: 'delivery_get', methods: ['get'])]
     public function get(ManagerRegistry $doctrine, int $id, Request $request): JsonResponse
     {
+        if (!$this->authentificationMiddleware->verify($request)) {
+            return $this->json([
+                'message' => 'You are not authentified or doesn\'t have the right to access this page'
+            ], Response::HTTP_UNAUTHORIZED);
+        }
         $delivery = $doctrine->getRepository(Delivery::class)->find($id);
         if (empty($delivery)) {
             return new JsonResponse(['message' => 'Trajet non trouvé'], 404);
@@ -151,8 +166,13 @@ class DeliveryController extends AbstractController
     }
 
     #[Route('/get', name: 'delivery_get_all', methods: ['get'])]
-    public function getAll(ManagerRegistry $doctrine): JsonResponse
+    public function getAll(ManagerRegistry $doctrine, Request $request): JsonResponse
     {
+        if (!$this->authentificationMiddleware->checkIfUserOffice($request)) {
+            return $this->json([
+                'message' => 'You are not authentified or doesn\'t have the right to access this page'
+            ], Response::HTTP_UNAUTHORIZED);
+        }
         $deliveries = $doctrine->getRepository(Delivery::class)->findAll();
         if (empty($deliveries)) {
             return new JsonResponse(['message' => 'Aucun trajet trouvé'], 404);
@@ -167,6 +187,11 @@ class DeliveryController extends AbstractController
     #[Route('/delete/{id}', name: 'delivery_delete', methods: ['delete'])]
     public function delete(ManagerRegistry $doctrine, int $id, Request $request): JsonResponse
     {
+        if (!$this->authentificationMiddleware->verify($request)) {
+            return $this->json([
+                'message' => 'You are not authentified or doesn\'t have the right to access this page'
+            ], Response::HTTP_UNAUTHORIZED);
+        }
         $delivery = $doctrine->getRepository(Delivery::class)->find($id);
         $entityManager = $doctrine->getManager();
         $entityManager->remove($delivery);
@@ -181,6 +206,11 @@ class DeliveryController extends AbstractController
     #[Route('/end/{id}', name: 'delivery_end', methods: ['put'])]
     public function endDelivery(ManagerRegistry $doctrine, int $id, Request $request): JsonResponse
     {
+        if (!$this->authentificationMiddleware->checkIfUserDriver($request)) {
+            return $this->json([
+                'message' => 'You are not authentified or doesn\'t have the right to access this page'
+            ], Response::HTTP_UNAUTHORIZED);
+        }
         $delivery = $doctrine->getRepository(Delivery::class)->find($id);
         $entityManager = $doctrine->getManager();
         $delivery->setEndDate(new DateTime());
