+++ back-end/src/Controller/VehicleController.php
@@ -5,17 +5,28 @@ namespace App\Controller;
 use App\Entity\User;
 use App\Entity\Vehicle;
 use App\Entity\VehicleType;
+use App\Middleware\AuthentificationMiddleware;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 class VehicleController extends AbstractController
 {
+    public function __construct(private AuthentificationMiddleware $authentificationMiddleware)
+    {
+    }
+
     #[Route('/add', name: 'vehicle_get', methods: ['post'])]
     public function add(ManagerRegistry $doctrine, Request $request): JsonResponse
     {
+        if (!$this->authentificationMiddleware->checkIfUserDriver($request)) {
+            return $this->json([
+                'message' => 'You are not authentified or doesn\'t have the right to access this page'
+            ], Response::HTTP_UNAUTHORIZED);
+        }
         $vehicle = new Vehicle();
         $datas = $request->getContent();
         if (!empty($datas['user_id']) && !empty($datas['vehicle_type_id'])) {
@@ -38,6 +49,11 @@ class VehicleController extends AbstractController
     #[Route('/get/{id}', name: 'vehicle_get', methods: ['get'])]
     public function get(ManagerRegistry $doctrine, int $id, Request $request): JsonResponse
     {
+        if (!$this->authentificationMiddleware->verify($request)) {
+            return $this->json([
+                'message' => 'You are not authentified or doesn\'t have the right to access this page'
+            ], Response::HTTP_UNAUTHORIZED);
+        }
         $vehicle = $doctrine->getRepository(Vehicle::class)->find($id);
         if (empty($vehicle)) {
             return new JsonResponse(['message' => 'Véhicule non trouvé'], 404);
@@ -46,8 +62,13 @@ class VehicleController extends AbstractController
     }
 
     #[Route('/get', name: 'vehicle_get_all', methods: ['get'])]
-    public function getAll(ManagerRegistry $doctrine): JsonResponse
+    public function getAll(ManagerRegistry $doctrine, Request $request): JsonResponse
     {
+        if (!$this->authentificationMiddleware->verify($request)) {
+            return $this->json([
+                'message' => 'You are not authentified or doesn\'t have the right to access this page'
+            ], Response::HTTP_UNAUTHORIZED);
+        }
         $vehicles = $doctrine->getRepository(Vehicle::class)->findAll();
         if (empty($vehicles)) {
             return new JsonResponse(['message' => 'Aucun véhicule trouvé'], 404);
@@ -62,6 +83,11 @@ class VehicleController extends AbstractController
     #[Route('/delete/{id}', name: 'vehicle_delete', methods: ['delete'])]
     public function delete(ManagerRegistry $doctrine, int $id, Request $request): JsonResponse
     {
+        if (!$this->authentificationMiddleware->checkIfUserOffice($request)) {
+            return $this->json([
+                'message' => 'You are not authentified or doesn\'t have the right to access this page'
+            ], Response::HTTP_UNAUTHORIZED);
+        }
         $vehicle = $doctrine->getRepository(Vehicle::class)->find($id);
         $entityManager = $doctrine->getManager();
         $entityManager->remove($vehicle);
